added messageDelete module to Index.js
Index.js got updated and now it makes an announcement on a pre-specified channel when someone deletes a message. channel can be specified with var "DELCHAN_ID" on config/env file.

Ban module has been re-reactivated after verifying that it's safe to use and it can only be used by trusted users (aka: users with admin permission and/or ban permission)